requirements.txt
Flask==2.0.1
numpy==1.21.0
tensorflow==2.5.0
joblib==1.0.1
matplotlib==3.4.2
psycopg2-binary==2.9.1
SQLAlchemy==1.4.21
